feat: don't use the scheme for manual_auth
the point of manual_auth is to be used as a fallback when either the server or the app fails to handle the tooth:// scheme. Having switching to manual_auth also depend on it is counterproductive.

Handling the StatusPage's description label's activate-link signal would need some workarounds so instead I turned into a separate label.