Composer dialog: set insensitive when sending
Currently when sending a post, the dialog does not change. This allows
the user to accidentally press "Publish" or use one of the accelerators
to send the post again. It's also just not good feedback on whether the
post is sending.

This commit changes the dialog such that once the publish process
begins, the "Publish" button, composer pages (edit/attach/poll), and
on_commit () callback are all disabled until the commit transaction
fails.