chore: downgrade logging from message to debug
These messages are very verbose (possibly many messages per second), yet
not interesting to a regular user, they're only for developers debugging
the app. Yet, they were logged to the system journal at the NOTICE
level, which is higher (numerically smaller) than even INFO, and so were
highligheted as important in the journalctl output.

Downgrade the logging to use debug (), which results in the messages not
being logged to the journal at all by default. It is still possible to
see the debugging messages by setting G_MESSAGES_DEBUG=Tuba (or "all").