feat(Status): ActionRow
let's decouple the actions from the status widget. Also get rid of LockableToggleButton. A toggle button wasn't the right widget for the job, we need to have control over the active state of the button but also handle clicked events ourselves. Network requests happen on click and the active state of the buttons can change multiple times (changes at the beginning and resets on error) before it finishes. Toggle button's toggle signal emits every time the active state changes which the locks had to take care of to avoid doing network requests when in the middle of one or during binding. Additionally, it's better to treat each button as unique instead of assuming they do the same thing but with a different entpoint. They are similar but not identical, boost will have different options eventually (visibility, quote), reply triggers a dialog and bookmark doesn't have a counter