feat: partial cleanup (#424)
* feat(BookWyrmPage): use builder files

* feat(Background): use cover instead

* feat(Avatar): do not store account or use signals

we only use the account to get the name and the avatar, no point in storing it

* chore(Timelines): use construct

* feat(Main): do not use lambdas for search button

* fix(LWW): infinite loop, statements

* feat(Base): do not use signals for shown status

we can use the setter instead

* chore(Base): move badge properties to Timeline

only used there

* fix(Status): remove menu model on destroy

* chore(EmojiLabel): remove namespaces

* chore: format

* feat(ProfileEdit): use Builder file

* fix(profile_edit): overlay expanding

* feat!: remove GTK < 4.10 compat

* feat!: remove GTK < 4.8 compat

* feat!: remove building without GtkSourceView

* chore(meson): bump gtk version

* feat(Lists): Builder files

* chore: lint

* revert: "chore(Base): move badge properties to Timeline"

This reverts commit 585d045174b6073e56b2fea166a6e3f1a00761ac.

* feat(PreviewCard): simplify, builder files

* feat: do not use 'using'

assuming the namespace is difficult for newcomers as one needs to have experience not only in all the namespaces but also the entire codebase

* feat: move more timelines to construct

* fix(FollowRequests): url <-> icon

* feat(Status): move reactions to a sep widget

* feat(Status): move follow request buttons to a sep widget

* fix(Notificatons): attention badge

* chore: lint

* feat(VoteBox): don't use lambdas

* feat(VoteBox): cleanup

* feat(LWW): avoid calling get_text on every allocate_shapes

label's text only gets updated when the label itself does, no reason to call it every time

* feat(LWW): avoid allocating children if there are none

* feat(LWW): avoid keeping track of 3 different arrays for children

instead use a struct

* chore: lint

* feat(Status): split actions

using a generic 'action' request for fav, unfav, boost, unboost, bookmark, unbookmark is kind-of limiting already but could break at a feature mastoapi change. It's better to treat each one as its own request as boosting already takes an extra optional param for visibility. The old 'action' function is still used internally

* feat(EntityCache): force insert option

this allows us to force update entities when we need to update them

* chore: construct object

* feat(Status): ActionRow

let's decouple the actions from the status widget. Also get rid of LockableToggleButton. A toggle button wasn't the right widget for the job, we need to have control over the active state of the button but also handle clicked events ourselves. Network requests happen on click and the active state of the buttons can change multiple times (changes at the beginning and resets on error) before it finishes. Toggle button's toggle signal emits every time the active state changes which the locks had to take care of to avoid doing network requests when in the middle of one or during binding. Additionally, it's better to treat each button as unique instead of assuming they do the same thing but with a different entpoint. They are similar but not identical, boost will have different options eventually (visibility, quote), reply triggers a dialog and bookmark doesn't have a counter

* fix(ActionsRow): unbind on destruction

* fix: only set paintable on is_loaded

* fix(ImageCache): avoid extra var for pixbuf

* feat(Settings): use array for key inits

* fix(Status): avoid lambda

* feat(Status): remove check_actions

this is something notifications should handle

* fix(Status): date not getting formatted correctly on patch when expanded

* feat(Status): do not disconnect the kind avi clicked signal

if the avi doesn't exist so does the signal

* feat(Status): make is_owned () and has_media () getters

* feat(Status): is_quote use array for the widgets

this allows us to skip some repetitive null checks and visibility toggles

* feat(Status): remove direct visibility indicator on bind if not applicable

for clearing before re-binding

* feat(Status): disconnect signals on bind

for clearing before re-binding

* feat(Status): remove obsolete actions box

* feat(Status): remove actions on rebind

* feat(VoteBox): simplify poll binding

BIDIRECTIONAL is enough

* feat(Status): remove attachments on rebind

* chore: lint

* feat(HtmlUtils): use libxml2

warning: libxml2 does not support HTML5 yet

* feat(tests)[HtmlUtils]: remove_tags

* chore: remove comment

* feat(Attachment.Item): avoid lambda

* fix(Status): open formal account directly

using on_activate_link does an extra request, this should be faster as it avoids resolving the account

* feat(NewAccount): hide next button on auto_auth

it's useless then, it's only used on manual auth

* feat(Conversation): remove useless on_error handler

this used to be needed to avoid segfaults, it has been fixed however

* feat: make use of RichLabel's mentions property

this skips resolving the account, should be faster

* feat(RichLabel): stop resolving if mention found

* chore: lint

* feat(Locales): optimize

ListStore needs objects so we can't create a const array of structs

instead of converting a Locale[] to Gee.ArrayList<Locale> on start and then appending each one individually to a liststore every time we need to use them, create a list store of all of them (using splice) on construct and call construct on boot

* fix(MediaViewer): change playing status only when item is done

* feat(MediaViewer): zoom_changed signal should be per item

instead of passing the whole media viewer to every item

* chore(MainWindow): remove unused const

* chore: lint

* feat: initial ListView

* fix(Timeline): nuke entity caches on activation

cache can cause urls to be wrong - even when keys are the same

* fix(Status): unbind bindings

* fix: more listview regressions

* fix(ci): use master for flatpak and disable snap

snap doesn't have the latest runtime yet

* fix(ci): use gnome-nightly