fix(profile): split cover and add to model (#481)
* feat(Relationship): to_string

* feat(ContentBase): remove all items from model but first

* fix(RelationshipButton): action activation format

it should be null, not empty

* fix(Profile): split profile and add to model

while using the listview headings felt more strightforward, it has its issues due to the cover being complicated. Specifically, setting vadjustment to 0 wouldnt start at the top, toggling the scrolledwindow's sensitivity wouldnt always recover its style and other binding and styling related issues. This commit moves the cover into its own entity and introduces a ProfileAccount object that has the account and the relationship but also is a Widgetizable, allowing us to prepend it to the model and transform into a widget automatically

* chore: lint

* fix: make cover non activatable

* fix: style