fix: ignore network monitor's lies
the network_changed signal is extremely unreliable for our needs. It can be triggered multiple times while a connection is being established (or just my machine and network made that happen) and it might end up with false even though it's online.

Instead, when it gets triggered set a timeout and check in 5 seconds if its online manually. By doing that we avoid triggering it needlessly: true => false (offline modal) => true (refresh + destroy modals) in the span of 1 second is expensive and useless. It now becomes true => false (check in 5s) => true (will do nothing since the value didn't change from the original one in those 5s)