feat: improve accessibility (#951)
* feat(ProfileCover): aria labels for avi and cover

* feat(Status+): aria labels for name and handle

* fix(DateTime): null iso

* feat(Status+): aria labels and tooltips on date

* feat(Status+): aria labels for icons

* feat(LWW): accessible custom emojis

the label child has the emojis replaced with the obj character. We need the label aria to have the custom emoji names *exactly* at the place they are visually.

* feat(ProfileCover+): ngettext stat buttons/labels

1 posts => 1 post

* feat(Status): tooltip texts for avis

* fix(StatusActionButton): should act as toggle button aria-wise

* feat(StatusActionButton): aria label with ngettext and amount

* fix(cover): a11y prop => relation

* fix(votebox): dont show exp date if it doesnt exist

* feat(status): container aria label summary

* feat(markupview): wip aria

* feat(ReactionsRow): make buttons act as toggle buttons

and wrap it in a flowboxchild manually so we can make it unfocusable (prevents double focus)

* feat(ActionsRow): comment out aria label templates

* feat(Attachment.Item): cleanup

previously, it would use a box and some complex wildcard styling. Let's just use the same overlays and simplify the styling by keeping adwaita's colors as is

* feat(Base): scroll to top button tooltip text

* feat(AccountSwitcher): avi button tooltip text

* feat(Base): state accessibility

when loading set it to busy, when there's a message, announce it

* feat(Status): ngettext attachments aria label section

* fix(Status): use this.spoiler_text for spoiler section of aria

* feat(Status): include reactions in aria

* feat(Status): include translation service in aria

* feat(Status): aria label on revealed spoiler box

* feat(Status): aria label quoted posts

* feat(VoteBox): set aria label

* feat(RichLabel): forward accessible text from LWW

* feat(MarkupView): set aria label to its children

since all children are richlabels and we forward their accessible_text, markupview can stitch them together and set it as its own. Description has to be nulled to avoid LWW children's aria labels being used.

* fix(EmojiLabel): null value

* feat(Emoji): set role as IMG

* fix(CEC): double focus

* fix(LWW): inspector crash

gtk inspector would crash due to this inf loop. We never actually use the getter so it would only happen on inspector. This would prevent us from debugging *any* LWW widget

* feat(LWW): don't act as a label

* fix(PreviewCard): double focus on explore

* fix(PreviewCard): focus border radius on explore

* feat(Status): use avatar description for name label

since it opens the profile as well

* feat(Search): add initial placeholder text

so the sreen reader has some context, rather than just 'form'.

* feat(Announcements): aria labels

keyboard navigation also needed some fixing due to the name label being untargetable (which makes sense on status as they are buttons, but they are not buttons here). Reaction buttons also were made toggle-like to match the status ones

* feat(Account): aria labels

close to the profile covers

* feat(ProfileCover): aria labels