fix(mv): clapper: Remove unnecessary Gst element creation
Since Clapper patch version 0.6.1 this is fixed to be done automatically
in ClapperGtk video, so no need to create another object instance here.

Most distros have now updated to 0.6.1 version, Clapper in Tuba is not
used by default and even with 0.6.0 the only dowside of removing it would
be wrong audio pitch on non-1x speed, so not a major enough reason to do
a version check just for that.