fix(mv): clapper: GIFV handling fixes (#1075)
* fix(mv): clapper: Do not use MPRIS for GIFs

Tuba now tries to treat GIFV as GIFs and not videos. For this reason
and consistency with how other clients/browser work, there should not
be a MPRIS controls for GIF media.

* fix(mv): clapper: Remove unnecessary Gst element creation

Since Clapper patch version 0.6.1 this is fixed to be done automatically
in ClapperGtk video, so no need to create another object instance here.

Most distros have now updated to 0.6.1 version, Clapper in Tuba is not
used by default and even with 0.6.0 the only dowside of removing it would
be wrong audio pitch on non-1x speed, so not a major enough reason to do
a version check just for that.

* fix(mv): Do not set autoplay for GIFV

Tuba always set player to playing (both GtkVideo and Clapper) for any kind of video when ready, so autoplay here is redundant.

* fix(mv): clapper: Fix handling of multiple media in a single post

When going "stopped" -> "paused" player will already start caching/buffering
until it has enough data to ba able to go into "playing" immediately. This causes
a problem where multiple created video widgets for multiple GIFV media download
all of videos at once while user views just on a one of them.

Fix this misbehavior by not going into "paused" when we are "stopped" which
means we are at initial state.