feat: use meson features to control optional dependencies
Use the standard Meson `feature` options to control optional link-time
dependencies, as documented in: https://mesonbuild.com/Build-options.html

More specifically, replace the current boolean options that are already
present for some of the dependencies with `feature` type options,
and add options for the remaining optional dependencies.  These options
integrate with `dependency()` calls automatically, making the resulting
code simpler and more reliable.  This also makes it possible to disable
all the optional dependencies explicitly.

This fixes a bug that Meson would attempt to link to the dependencies
(such as `clapper`) even if they were disabled, since disabling
the options did not disable the dependency checks,
and the dependencies were used in the `final_deps` unconditionally.

I've left `libwebp` as-is, that is without an option, since it is only
used to emit a warning.  I have also removed a redundant `clapper`
version check, since the current dependency is higher than that.