fix(SandwitchSourceView): yoffset and size allocations
an issue with the prior implementation was that the children would be 'cut off' when updating. After a very lengthy debugging process I discovered the issue. The yoffset gets updated after setting the top/bottom margins when we call size_allocate on the textview, but we would only call it between measuring and allocating the top and bottom children, which ended up messing with both. The correct way to do this is: 1. measure top & set top margin, 2. measure bottom & set bottom margin, 3. allocate textview, 4. get the yoffset, 5. allocate the top and bottom widgets as we did prior