fix(editor): placeholder sizing
due to the way the textview measures and allocates the overlay widgets, it causes its own size to be incosistent. We use overlay widgets for the placeholder, which is a label, and we make it invisible when the textview is not empty. When that happens the textview size also changes and theres a jarring effect. If the label is set to wrapping, it has 0 width which fixes that but also doesn't extend horizontally. This commit adds a wrapper widget that sets the natural width to max(500, min), allowing the textview to treat it as 0 width min with wrapping while displaying it fully with a big nat width which also acts as a liberal way to display different translations that might be longer