fix(editor): force reallocate placeholder
in gtk < 4.19.1, textview placeholders do not stay in position while scrolling. This has been fixed but it might not make it in a 4.18 patch, so this is a hack around it by asking the placeholder's parent, which is a textview child, to reallocate when the editor allocates, forcing it to get in position.