fix(Drive): properly implement our custom scroll structure
previously we would replace the scrolled child, but that child and scrollwindow were used by Base and would leave broken accounting warnings (and I believe a segfault). Now we rename the page to old-content and add a new page with our structure named content, essentially replacing the Base content without removing it.